Latest Logstash 6.3 and Logstash Puppet module 6.3
This is a big refactor in order to move to the latest Logstash 6.3 and
also the latest Logstash Puppet module, also version 6.3.

Part of this change involves using the new Multiple Pipelines feature,
as suggested in https://github.com/elastic/puppet-logstash/issues/144

And because of how that works, it was necessary to remove the Shipper
and Indexer profiles, and have just a single Logstash profile. It can be
configured as either a Shipper, Indexer, or both, by passing in
different Configfiles and PatternFiles and Pipeline configs.