Add a sleep for logstash service
I am not at this point sure why, but it seems that the Logstash service
can take quite a while to start. To avoid Beaker tests failing due to a
slow-starting Logstash service, I add a pause here until the Logstash
service really starts.

Also switches from the use of Exec-based sleeps to the
basti1302-wait_for type, which makes it easy to sleep for a maximum
number of polling intervals.