Resolve cyclic dependency in firewall setup
Before this, a cyclic dependency was seen in the firewall module setup.
On investigation, I found that the setup instructions at the firewall
module project had changed and after the setup in here to match, the
cyclic dependency was resolved. I didn't spend any more time on it after
that.